+++ pkg/accessreview/drivers/google_analytics.go
@@ -406,3 +406,74 @@ func (r *googleAnalyticsNameResolver) ResolveInstanceName(ctx context.Context) (
 
 	return resp.DisplayName, nil
 }
+
+// ListGoogleAnalyticsOrganizations fetches the GA4 accounts the authenticated
+// Google user can access, surfacing each account's numeric ID as the picker
+// slug. Listing accounts requires the analytics.readonly scope.
+func ListGoogleAnalyticsOrganizations(ctx context.Context, httpClient *http.Client) ([]Organization, error) {
+	var orgs []Organization
+
+	pageToken := ""
+
+	for range maxPaginationPages {
+		endpoint, err := googleAnalyticsURL(pageToken, nil, "v1alpha", "accounts")
+		if err != nil {
+			return nil, err
+		}
+
+		req, err := http.NewRequestWithContext(ctx, http.MethodGet, endpoint, nil)
+		if err != nil {
+			return nil, fmt.Errorf("cannot create google analytics accounts request: %w", err)
+		}
+
+		req.Header.Set("Accept", "application/json")
+
+		resp, err := httpClient.Do(req)
+		if err != nil {
+			return nil, fmt.Errorf("cannot fetch google analytics accounts: %w", err)
+		}
+
+		if resp.StatusCode != http.StatusOK {
+			_ = resp.Body.Close()
+
+			return nil, fmt.Errorf("cannot fetch google analytics accounts: unexpected status %d", resp.StatusCode)
+		}
+
+		var out struct {
+			Accounts []struct {
+				Name        string `json:"name"`
+				DisplayName string `json:"displayName"`
+			} `json:"accounts"`
+			NextPageToken string `json:"nextPageToken"`
+		}
+
+		decodeErr := json.NewDecoder(resp.Body).Decode(&out)
+		_ = resp.Body.Close()
+
+		if decodeErr != nil {
+			return nil, fmt.Errorf("cannot decode google analytics accounts response: %w", decodeErr)
+		}
+
+		for _, a := range out.Accounts {
+			id := strings.TrimPrefix(a.Name, "accounts/")
+			if id == "" {
+				continue
+			}
+
+			displayName := a.DisplayName
+			if displayName == "" {
+				displayName = id
+			}
+
+			orgs = append(orgs, Organization{Slug: id, DisplayName: displayName})
+		}
+
+		if out.NextPageToken == "" {
+			return orgs, nil
+		}
+
+		pageToken = out.NextPageToken
+	}
+
+	return nil, fmt.Errorf("cannot list all google analytics accounts: %w", ErrPaginationLimitReached)
+}
